refactor(type): extract TickOptions from BaseOptions (#78)
* refactor(type): extract TickOptions from BaseOptions due to not all scales supports getTicks()

* test(tick-methods): add test suite for basic tick method
