+++ __tests__/unit/scales/base.spec.ts
@@ -1,6 +1,5 @@
 import { Base } from '../../../src/scales/base';
 import { BaseOptions, Domain, Range } from '../../../src/types';
-import { ticks } from '../../../src/tick-method/basic';
 
 class Scale extends Base<BaseOptions> {
   // eslint-disable-next-line class-methods-use-this
@@ -22,28 +21,20 @@ describe('Scale', () => {
   test('Scale() has expected defaults', () => {
     const s = new Scale();
     // @ts-ignore
-    const { formatter, tickMethod, ...restOptions } = s.options;
+    const { formatter, ...restOptions } = s.options;
 
     expect(restOptions).toEqual({
       domain: [0, 1],
       range: [0, 1],
-      tickCount: 5,
-      tickInterval: 10,
     });
     expect(formatter(1)).toBe('1');
-    // @ts-ignore
-    expect(tickMethod(s.options)).toEqual([]);
-    expect(tickMethod).toEqual(ticks);
   });
 
   test('Scale(options) override defaults', () => {
     const s = new Scale({
-      tickCount: 20,
       domain: [0, 10],
     });
 
-    // @ts-ignore
-    expect(s.options.tickCount).toBe(20);
     // @ts-ignore
     expect(s.options.domain).toEqual([0, 10]);
   });
@@ -65,28 +56,11 @@ describe('Scale', () => {
     const s = new Scale();
 
     s.update({
-      tickCount: 10,
       domain: [0, 10],
     });
 
     const options = s.getOptions();
-
-    expect(options.tickCount).toBe(10);
     expect(options.domain).toEqual([0, 10]);
-  });
-
-  test('getTicks() call options.tickMethod and return its return value', () => {
-    const s = new Scale();
-    const mockFn = jest.fn();
-    s.update({
-      tickMethod: () => {
-        mockFn();
-        return [1, 2, 3, 4, 5];
-      },
-    });
-
-    expect(s.getTicks()).toEqual([1, 2, 3, 4, 5]);
-    expect(mockFn).toBeCalled();
   });
 
   test('clone() return a scale belong to same class', () => {
@@ -106,13 +80,13 @@ describe('Scale', () => {
     const s1 = s.clone();
 
     s.update({
-      tickCount: 20,
+      domain: [0, 10],
     });
-    expect(s1.getOptions().tickCount).toBe(5);
+    expect(s1.getOptions().domain).toEqual([0, 1]);
 
     s1.update({
-      tickCount: 30,
+      domain: [0, 100],
     });
-    expect(s.getOptions().tickCount).toBe(20);
+    expect(s.getOptions().domain).toEqual([0, 10]);
   });
 });
