+++ src/scales/constant.ts
@@ -1,5 +1,6 @@
 import { ConstantOptions, Domain, Range } from '../types';
 import { Base } from './base';
+import { ticks } from '../tick-method/basic';
 
 export class Constant extends Base<ConstantOptions> {
   /**
@@ -7,7 +8,12 @@ export class Constant extends Base<ConstantOptions> {
    * @param options 需要自定义配置的选项
    */
   constructor(options?: Partial<ConstantOptions>) {
-    super(options, { range: [0] });
+    super(options, {
+      range: [0],
+      tickCount: 5,
+      tickInterval: 10,
+      tickMethod: ticks,
+    });
   }
 
   /**
@@ -37,5 +43,13 @@ export class Constant extends Base<ConstantOptions> {
    */
   public clone() {
     return new Constant(this.options);
+  }
+
+  /**
+   * 根据比例尺的配置去生成 ticks，该 ticks 主要用于生成坐标轴
+   * @returns 返回一个 ticks 的数组
+   */
+  public getTicks(): Range<ConstantOptions>[] {
+    return this.options.tickMethod(this.options);
   }
 }
