+++ src/types.ts
@@ -6,7 +6,7 @@ export type TickMethod<T> = (options?: T) => any[];
  * R：值域元素的类型
  * T：tickMethod 配置项的类型
  */
-export type BaseOptions<D = any, R = any, T = any> = {
+export type BaseOptions<D = any, R = any> = {
   /** 当需要映射的值不合法的时候，返回的值 */
   unknown?: any;
   /** 值域，默认为 [0, 1] */
@@ -15,6 +15,12 @@ export type BaseOptions<D = any, R = any, T = any> = {
   domain?: D[];
   /** tick 格式化函数，会影响数据在坐标轴 axis、legend、tooltip 上的显示 */
   formatter?: (x: R) => string;
+};
+
+/**
+ * 支持 getTicks 的比例尺的选项
+ */
+export type TickOptions<T = any> = {
   /** tick 个数，默认值为 5 */
   tickCount?: number;
   /** tick 间隔的最大值，默认值为 10 */
@@ -32,6 +38,6 @@ export type Range<O extends BaseOptions> = O['range'][number];
 /** 获得比例尺选项中 unknown 的类型 */
 export type Unknown<O extends BaseOptions> = O['unknown'];
 
-export type IdentityOptions = BaseOptions<number, number>;
+export type IdentityOptions = BaseOptions<number, number> & TickOptions;
 
-export type ConstantOptions = BaseOptions<number | string, number | string>;
+export type ConstantOptions = BaseOptions<number | string, number | string> & TickOptions;
